% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{synAncestry}
\alias{synAncestry}
\title{Retrieve parentId chain}
\usage{
synAncestry(...)
}
\arguments{
\item{...}{One or more Synapse IDs for which ancestry is to be reconstructed}
}
\value{
For each requested ID, the function returns a vector of parentId ancestry.
The first entry in each vector is the query ID itself; the last entry is the project.
}
\description{
Recursively reconstructs parentId ancestry for a given Synapse ID
}
\examples{
\dontrun{
synAncestry( "syn15663039", "syn1695362" )
# $syn15663039
# [1] "syn15663039" "syn15673834" "syn15673837" "syn12180284"
#
# $syn1695362
# [1] "syn1695362" "syn1695324" "syn2812925" "syn300013"
}
}
