% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synChildren}
\alias{synChildren}
\title{Retrieve child entities}
\usage{
synChildren(..., .fullInfo = FALSE)
}
\arguments{
\item{...}{One or more Synapse IDs for which children should be retrieved}

\item{.fullInfo}{set to TRUE to retrieve the full range of information for each child (default: FALSE)}
}
\value{
A vector of child instance IDs if .fullInfo is FALSE. A data frame with all relevant info, if .fullInfo is TRUE.
}
\description{
A wrapper around synGetChildren() for multiple Synapse IDs
}
