% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updown.R
\name{synDownloader}
\alias{synDownloader}
\title{Instantiate a Synapse downloader}
\usage{
synDownloader(dloc, useCache = FALSE, ...)
}
\arguments{
\item{dloc}{Path to a local folder where downloaded files will be stored}

\item{useCache}{Use locally cached file if file has already been download and MD5 hasn't changed}

\item{...}{Additional arguments to be passed to synapser::synGet()}
}
\value{
A downloader function that recognizes synapse IDs and downloads the associated files.
}
\description{
Function creates and returns a downloader for files on Synapse.
}
\details{
A downloader is itself a function that accepts a vector or list of character strings.
Any recognized Synapse IDs get downloaded to the local folder specified by dloc.
The downloader then returns local paths to these files. Strings not recognized to be
valid Synapse IDs are returned as is.
}
\examples{
\dontrun{
syn <- synDownloader( "/data", ifcollision="overwrite.local" )
syn( "syn15663039", "localFile.csv" )
# [1] "/data/mtcars.csv" "localFile.csv"
}
}
