% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{synGlob}
\alias{synGlob}
\title{Traverse a synapse path with wildcard matching}
\usage{
synGlob(sid, ..., .nested = FALSE)
}
\arguments{
\item{sid}{Synapse ID of the starting entity}

\item{...}{One or more names for constructing a path. Accepts individual names, vectors and lists}

\item{.nested}{Return synapse paths as flat vector (FALSE) or as nested list
following the directory structure (TRUE)}
}
\value{
Synapse IDs of the "plucked" entities
}
\description{
Starting from the provided synapse ID, traverses descendants by name
that can contain wildcards (*) that match an arbitrary number of characters.
}
\details{
Details about glob wildcard character matching are available in the
\href{https://en.wikipedia.org/wiki/Glob_(programming)}{Wikipedia article}.
}
\examples{
\dontrun{
synGlob( "syn26348949", "Runs", "*", "meta", "*.csv" )
}
}
