% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synMkdir}
\alias{synMkdir}
\title{Create a synapse directory hierarchy}
\usage{
synMkdir(sid, ..., .recursive = FALSE)
}
\arguments{
\item{sid}{Synapse ID of the starting entity}

\item{...}{One or more names for constructing a directory hierarchy.
Accepts individual names, vectors and lists}

\item{.recursive}{Whether to create intermediate directories if they don't exist (default: FALSE)}
}
\value{
Synapse ID of the "plucked" entity
}
\description{
Starting from the provided synapse ID, traverse descendants by name like
\code{\link[=synPluck]{synPluck()}}. If a child doesn't exist, create a directory with that name.
If \code{.recursive} is set, also create intermediate directories as necessary.
}
\details{
Returns Synapse ID of final child like \code{\link[=synPluck]{synPluck()}} if the directory already
exists.
}
\examples{
\dontrun{
synMkdir( "syn1773110", "mRNA", "Counts", "htseq-count" )
# [1] "syn2822494"
}
}
