% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synName}
\alias{synName}
\title{Look up filename based on synapse ID}
\usage{
synName(...)
}
\arguments{
\item{...}{One or more Synapse IDs. Accepts individual IDs, vectors or lists.}
}
\value{
A character vector of filenames corresponding to the provided synapse IDs
}
\description{
Retrieves file names associated with given synapse IDs
}
\examples{
\dontrun{
synName( "syn1896429", "syn1901530" )
#                            syn1896429                              syn1901530
#        "ESTIMATE_scores_PANCAN11.tsv" "ESTIMATE_scores_PANCAN11_RNASeqV2.tsv"
}
}
