% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synParent}
\alias{synParent}
\title{Retrieve parentId}
\usage{
synParent(...)
}
\arguments{
\item{...}{One or more Synapse IDs for which parent IDs must be retrieved}
}
\value{
Named character vector that maps the input Synapse IDs their parentIds
}
\description{
Retrieves parentIds for one or more of the provided Synapse IDs
}
\examples{
\dontrun{
synParent("syn15663039", "syn1695362" )
#   syn15663039    syn1695362
# "syn15673834"  "syn1695324"
}
}
