% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synq}
\alias{synq}
\title{Streamlined version of synQuery()}
\usage{
synq(..., .echo = TRUE)
}
\arguments{
\item{...}{Arguments defining the query. Unnamed arguments make up the fields to be selected.
Named arguments define the query conditions.}

\item{.echo}{Whether to display the final query string. (Default: TRUE)}
}
\value{
The function always returns a data frame with columns defined by the unnamed arguments. If the query returned no results, the data frame will have zero rows (but non-zero number of columns).
}
\description{
An intuitive interface for composing Synapse queries
}
\examples{
\dontrun{
synq( "id", "name", "projectId", parentId = "syn15673834", type = "dataset" )
# select id,name,projectId from entity where "parentId"=="syn15673834" and "type"=="dataset" 
# # A tibble: 1 x 3
#   projectId id          name      
#       <dbl> <chr>       <chr>     
# 1  12180284 syn15663039 mtcars.csv

synq( "id", "name", projectId = "12180284", type="Nonexistent", .echo=FALSE )
# # A tibble: 0 x 2
# # ... with 2 variables: id <chr>, name <chr>

}
}
