% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synRename}
\alias{synRename}
\title{Rename a synapse entity}
\usage{
synRename(sid, newName)
}
\arguments{
\item{sid}{Synapse ID of the entity to rename}

\item{newName}{New filename that should be assigned to the entity}
}
\value{
Synapse ID of the entity (for integration with the %>% pipe)
}
\description{
Assigns a new file name to the synapse entity identified by its ID
}
