% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synSize}
\alias{synSize}
\title{Retrieve the size of a synapse entity}
\usage{
synSize(...)
}
\arguments{
\item{...}{One or more Synapse IDs for which parent IDs must be retrieved}
}
\value{
Named character vector that maps the input Synapse IDs their file sizes
}
\description{
Returns the size of a file. For folders, returns the sum of sizes for all files inside that folder
}
\examples{
\dontrun{
synSize( "syn1695362", "syn1695324" )
# syn1695362  syn1695324
# 1678169935 11682238368
}
}
