% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synUpload}
\alias{synUpload}
\title{Upload and annotate a file}
\usage{
synUpload(.fn, .dest, ..., .forceVer = TRUE)
}
\arguments{
\item{.fn}{Filename of the local file to upload}

\item{.dest}{Synapse ID of the destination Folder}

\item{...}{Named arguments specifying key-value pairs to serve as annotations}

\item{.forceVer}{Maps to forceVersion argument of synStore(), which is defined as
"Indicates whether the method should increment the version of the object even if
nothing has changed. Defaults to TRUE."}
}
\value{
Annotated synapse entity object
}
\description{
The function uploads a local file and annotates it with fields provided as named arguments
}
\examples{
\dontrun{
synUpload( "mtcars.csv", "syn12180284", Field1 = c("abc","def"), Field2 = 123 )
}
}
