% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{isSynID}
\alias{isSynID}
\title{Identifies if its argument is a valid Synapse ID}
\usage{
isSynID(ids)
}
\arguments{
\item{ids}{A character vector containing one or more strings to identify.}
}
\value{
Logical values indicating whether each element of the input vector is a valid Synapse ID.
}
\description{
Identifies if its argument is a valid Synapse ID
}
\examples{
isSynID( c("syn1234", "syn", "syn11ab") )
# [1]  TRUE FALSE FALSE
}
