% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{isSynID}
\alias{isSynID}
\title{Check validity of Synapse IDs}
\usage{
isSynID(...)
}
\arguments{
\item{...}{One or more Synapse IDs. Accepts individual IDs, vectors or lists.}
}
\value{
Logical values indicating whether each element of the input is a valid Synapse ID.
}
\description{
Identifies if its argument is a valid Synapse ID character string
}
\examples{
isSynID( "syn1234", "syn", "syn123ab" )
# [1]  TRUE FALSE FALSE
isSynID( list( mtcars, 123, "syn123" ) )
# [1] FALSE FALSE  TRUE
}
