% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{synStoreMany}
\alias{synStoreMany}
\title{Upload a directory structure to Synapse}
\usage{
synStoreMany(localPath, parentId)
}
\arguments{
\item{localPath}{A character string to the local file / directory to upload}

\item{parentId}{A synapse ID of the hosting project or parent folder on Synapse}
}
\value{
A nested list capturing the created hierarchy of Synapse IDs
}
\description{
Recursively uploads a directory and all of its content to the given Synapse project/folder
}
